requirements.txt
flask==3.0.0
flask-cors==4.0.0
numpy>=1.24.4
pillow>=10.1.0
librosa>=0.10.1
opencv-python>=4.8.1.78
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.35.0
exifread>=3.0.0
scipy>=1.11.4
werkzeug>=3.0.1
